pyproject.toml
[tool.poetry]
name = "grpc-whisper"
version = "0.1.0"
description = "gRPC server for OpenAI's Whisper Models"
authors = ["Fariz Tumbuan <fariz.tumbuan@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
openai-whisper = "^20230314"
grpcio = "^1.54.0"
grpcio-tools = "^1.54.0"
stable-ts = "^2.5.3"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
click = "^8.1.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
